build: factor reusable buildExtension out of build.js
